refactor(richtext): swap regex pipeline for turndown + DOMPurify
Complete the swap from the regex-based converter by extracting the four
project-specific turndown overrides (compact list items, GFM-style
double-tilde strikethrough, `<u>` → italic, language-hinted code
fences) into a dedicated `turndownRules.ts` helper. The main
`htmlToMarkdown.ts` file now reads top-to-bottom as:

  1. ALLOWED_TAGS / ALLOWED_ATTR / ALLOWED_URI_REGEXP for DOMPurify
  2. TurndownService instantiation + `.use(gfm)` + registerCustomRules
  3. The public `htmlToMarkdown` function (sanitise → convert →
     whitespace normalise)

No behavioural change — the 1688-test suite, including the original
htmlToMarkdown tests, the WS-A XSS suite, and the new GFM parity suite
from the previous commit, continues to pass unchanged.
